use size_t for datatype
git-svn-id: http://svn.xp-dev.com/svn/wxextension@2422 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/src/textfile.cpp
@@ -336,7 +336,7 @@ bool wxExTextFile::Parse()
     }
   }
 
-  for (int i = 0; i < GetLineCount() && !Cancelled() && !m_FinishedAction; i++)
+  for (size_t i = 0; i < GetLineCount() && !Cancelled() && !m_FinishedAction; i++)
   {
     wxString& line = GetLine(i);
